Extract form to a new component
