+++ src/features/auth/components/Input.tsx
@@ -1,13 +1,14 @@
-import { type ChangeEvent, type ReactElement, type RefObject } from "react";
+import { type ChangeEvent, type ReactElement } from "react";
 
 interface InputProps {
   label: string;
   name: string;
-  type: "text" | "email" | "password" | "submit";
+  type: "text" | "email" | "password";
   value: string;
   onChange?: (event: ChangeEvent<HTMLInputElement>) => void;
   autoFocus: boolean;
   disabled: boolean;
+  autoComplete?: string;
 }
 
 export const Input = ({
@@ -18,29 +19,26 @@ export const Input = ({
   onChange,
   autoFocus,
   disabled,
+  autoComplete,
 }: InputProps): ReactElement => {
   const id = `${name.toLowerCase().replace(/\s+/g, "-")}`;
 
-  if (type === "submit") {
-    return <input id={id} type={type} value={label} disabled={disabled} />;
-  } else {
-    return (
-      <>
-        <label>
-          {label}
-          <input
-            id={id}
-            type={type}
-            name={name}
-            value={value}
-            onChange={onChange}
-            autoComplete="on"
-            required
-            autoFocus={autoFocus}
-            disabled={disabled}
-          />
-        </label>
-      </>
-    );
-  }
+  return (
+    <>
+      <label>
+        {label}
+        <input
+          id={id}
+          type={type}
+          name={name}
+          value={value}
+          onChange={onChange}
+          autoComplete={autoComplete}
+          required
+          autoFocus={autoFocus}
+          disabled={disabled}
+        />
+      </label>
+    </>
+  );
 };
