+++ src/features/auth/css/Login.css
@@ -23,7 +23,10 @@
   width: 400px;
 }
 
-#login-page .login-form input {
+#login-page .login-form input,
+#login-page .login-form button {
+  border: 1px solid #ccc;
+  border-radius: 6px;
   font-size: 20px;
   margin-top: 0.5rem;
   margin-bottom: 1.5rem;
@@ -40,18 +43,14 @@
   padding-bottom: 0.8rem;
 }
 
-#login-page .login-form input {
-  border: 1px solid #ccc;
-  border-radius: 6px;
-}
-
-#login-page .login-form input[type="submit"] {
+#login-page .login-form button {
   background-color: #3182ce;
   color: whitesmoke;
   padding: 0.8rem;
+  cursor: pointer;
 }
 
-#login-page .login-form input[type="submit"]:disabled {
+#login-page .login-form button:disabled {
   background-color: #3182ce;
   cursor: not-allowed;
   opacity: 0.7;
